+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# For next release
+  * **Raphael Pour**
+    * metadata: render posts with date instead of full timestamp
+
+*Not released yet*
+
 # Minor Release v2.1.0 (2024-01-15)
   * **Raphael Pour**
     * templates: